+++ perl/FSM/Composition/Net.pm
@@ -11,7 +11,7 @@ sub new ($class, %args) {
         name => $args{name},
         width => $args{width} // 1,
         source => $args{source},
-        targets => $args{targets} || [],
+        targets => [@{$args{targets} || []}],
         declaration_keyword => $args{declaration_keyword},
         signed => ($args{signed} // 0) ? 1 : 0,
         state_model => $args{state_model},
@@ -23,12 +23,20 @@ sub new ($class, %args) {
 sub name ($self) { return $self->{name} }
 sub width ($self) { return $self->{width} }
 sub source ($self) { return $self->{source} }
-sub targets ($self) { return $self->{targets} }
+sub targets ($self) { return [@{$self->{targets} || []}] }
 sub declaration_keyword ($self) { return $self->{declaration_keyword} }
 sub signed ($self) { return ($self->{signed} // 0) ? 1 : 0 }
 sub state_model ($self) { return $self->{state_model} }
 sub declared_type_name ($self) { return $self->{declared_type_name} }
 sub declared_type_spec ($self) { return _clone_structured_value($self->{declared_type_spec}) }
+
+sub add_target ($self, $target) {
+    return $self->targets
+        unless defined($target) && length($target);
+
+    push @{$self->{targets}}, $target;
+    return $self->targets;
+}
 
 sub _clone_structured_value ($value) {
     return undef unless defined $value;